@@ -1,34 +1,24 @@
 module.exports = {
   env: {
-    browser: true,
-    commonjs: true,
-    es6: true,
     node: true,
+    browser: true,
+    es2021: true,
     'jest/globals': true
   },
-  plugins: [
-    'jest'
-  ],
   extends: [
     'standard',
     'plugin:jest/recommended'
   ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly'
-  },
+  parser: '@typescript-eslint/parser',
   parserOptions: {
-    ecmaVersion: 2018
+    ecmaVersion: 12,
+    sourceType: 'module'
   },
+  plugins: [
+    '@typescript-eslint',
+    'jest'
+  ],
   rules: {
-    indent: [
-      'error',
-      2,
-      {
-        SwitchCase: 1,
-        VariableDeclarator: 'first'
-      }
-    ],
     'jest/no-disabled-tests': 'warn',
     'jest/no-focused-tests': 'error',
     'jest/no-identical-title': 'error',
